requirements.txt
praat-parselmouth>=0.3.3
matplotlib>=3.9.0
numpy>=1.21.0
pandas>=1.3.0
